Changed the mcuconf.h to match the STM32F4 Discovery board.
Put all of the usb_cdc in it's own .h file (to clean up main.c)

Removed web/ since this is only a FatFS test.

Enabled STRFUNC and MKFS in ffconf.h so that string functions are enabled.

Edded EMBEST_DMSTF4BB board files to this repository.

Removed references to the webserver in main.

Added functions to make directory, create a 'hello.txt' file and to echo back any text file with 'cat'.